+++ src/pages/tabs/tabs.html
@@ -1,4 +1,4 @@
-<ion-tabs >
+<ion-tabs selectedIndex="1">
   <ion-tab [root]="tab1Root" tabTitle="Favorites" tabIcon="star-outline"></ion-tab>
   <ion-tab [root]="tab2Root" tabTitle="Wheel" tabIcon="aperture"></ion-tab>
   <ion-tab [root]="tab3Root" tabTitle="Restauraunts" tabIcon="restaurant"></ion-tab>
