+++ src/pages/wheel/wheel.scss
@@ -1,3 +1,8 @@
 page-wheel {
-
+  .background{
+    background-color:  #566573;
+    margin-left: auto;
+    margin-right: auto;
+    display: block;
+  }
 }
